server/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.5
pillow==11.0.0
PyJWT==2.9.0
pyotp==2.9.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
